requirements.txt
torch>=2.2
torchvision>=0.17
numpy>=1.23
Pillow>=10.0
opencv-python>=4.8
tqdm>=4.66
PyYAML>=6.0
